Separate caller into scripts folder and rest into package darepy_sans
