+++ scripts/caller_merging.py
@@ -9,11 +9,9 @@ Created on Thu Aug 17 16:33:06 2023
 # %% Plot the detector distances in the same graphic
 path_dir = "C:/Users/lutzbueno_v/Documents/Analysis/data/SANS-LLB/2024_SANS-LLB/DarePy-SANS/"
 
-
 import os
-os.chdir(path_dir + '/codes/')
-path_dir_an = path_dir + '/analysis/'
-import post_processing as pp
+path_dir_an = os.path.join(path_dir, 'analysis')
+import darepy_sans.post_processing as pp
 
 # %% STEP 1: PLOT DATA TOGETHER
